% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{create_flag_threshold}
\alias{create_flag_threshold}
\title{create_flag_threshold}
\usage{
create_flag_threshold(data, var, threshold_data)
}
\arguments{
\item{data}{...}

\item{vars3}{...}
}
\value{
...
}
\description{
...
}
\examples{

}
