% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flag_midpoints}
\alias{flag_midpoints}
\title{flag_midpoints}
\usage{
flag_midpoints(data, vars, midpoint = 3, ratio = 0.5)
}
\arguments{
\item{data}{A tibble (or a compatible data structure)}

\item{vars}{Any expression that can be used with dplyr::select}

\item{midpoint}{An integer (if not provided, its default value is 3)}

\item{ratio}{Any number between 0 and 1 (if not provided, its default value is 0.5)}
}
\value{
...
}
\description{
...
}
