% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{flag_missing}
\alias{flag_missing}
\title{flag_missing}
\usage{
flag_missing(data, vars, ratio = 0.33)
}
\arguments{
\item{data}{A data set object}

\item{vars}{A data set object}

\item{ratio}{A data set object}
}
\value{
...
}
\description{
...
}
\examples{

}
