% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flag_straight}
\alias{flag_straight}
\title{flag_straight}
\usage{
flag_straight(data, vars)
}
\arguments{
\item{data}{A tibble (or a compatible data structure)}

\item{vars}{Any expression that can be used with dplyr::select
and it should correspond to the items that have been displayed
to respondents as a grid question}
}
\value{
A binary vector
}
\description{
flag_ straight uses two arguments: data, and vars and it creates a binary vector.
}
