% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{flag_times}
\alias{flag_times}
\title{flag_times}
\usage{
flag_times(data, threshold_file, ratio = 0.1)
}
\arguments{
\item{data}{A tibble}

\item{threshold_file}{The name of the file with variable names, number of characters and number of subquestions}

\item{ratio}{can by any number between 0 and 1}
}
\value{
...
}
\description{
...
}
