% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{issp2020}
\alias{issp2020}
\title{ISSP 2020}
\format{
A tibble: 179 × 118
}
\describe{}

\source{
\url{}
}
\usage{
issp2020
}
\description{
A dataset collected from a sample of students at Aristotle University of Thessaloniki using the 2020 questionnaire of the International Social Survey Programme.

}
\keyword{datasets}
