% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{load_dataset}
\alias{load_dataset}
\title{load_dataset}
\usage{
load_dataset()
}
\arguments{
\item{dataset}{A data set object}
}
\value{
A data frame object in the form of a list
}
\description{
Load internal dataset for testing
}
\examples{
load_dataset()
}
