% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{flag_midpoints}
\alias{flag_midpoints}
\title{flag_missing}
\usage{
flag_midpoints(data, vars2, midpoint = 3, ratio2 = 0.5)
}
\arguments{
\item{data}{A data set object}

\item{vars2}{A data set object}

\item{midpoint}{A data set object}

\item{ratio2}{A data set object}
}
\value{
...
}
\description{
...
}
\examples{

}
